Added Compositing Classes
Moved compositing effects (screen-space reflection and refraction
effects, bloom, and whatever is to come) to `Compositing.h` and
`Compositing.cpp`. New scratch buffers, descriptor sets, and pipelines
are set up, this commit is a semi-functional anchor point to rollback to
in case integration and removal of old system goes horribly wrong.